api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
elasticsearch
open-clip-torch
torch